rename gate from ggate
