+++ gate/client/image.go
