--- ggate/README.md
@@ -1,3 +1,0 @@
-# gGate
-
-This is gRPC Gateway.
